feat: ✨ Old version in new folder structure

+++ packages/discord.js/src/managers/ReactionManager.js
@@ -58,6 +58,7 @@ class ReactionManager extends CachedManager {
    * @returns {Promise<Message>}
    */
   async removeAll() {
+    if (this.message.deleted) return this.message;
     await this.client.api.channels(this.message.channelId).messages(this.message.id).reactions.delete();
     return this.message;
   }
